pyproject.toml
[project]
name = "food-analyzer-agent"
version = "3.0.0"
description = "Agente inteligente con DSPy para análisis de comida con visión GPT-4"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "motor>=3.7.1",
    "pymongo>=4.15.1",
    "uvicorn>=0.35.0",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "dspy-ai>=2.5.0",
]